requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
braintree==3.45.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-bootstrap4==1.1.1
django-cleanup==4.0.0
easy-thumbnails==2.7
idna==2.8
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
